Title: 사각형 채우기 3, Time: 116ms, Memory: 10MB, Status: Wrong Answer - CodeTree

+++ 240916/사각형 채우기 3/rectangle-fill-3.java
@@ -4,7 +4,7 @@ public class Main {
 
     public static int N;
 
-    public static int[] dp;
+    public static long[] dp;
 
     public static void main(String[] args) {
         // 여기에 코드를 작성해주세요.
@@ -13,7 +13,7 @@ public class Main {
 
         N = sc.nextInt();
 
-        dp = new int[1001];
+        dp = new long[1001];
 
         dp[1] = 2;
         dp[2] = 7;
